requirements.txt
matplotlib
opencv-python
pandas
numpy
ipywidgets
wandb
ultralytics
PyYAML
Pillow